sql-python-rdms/cmsc408-sp2025-hw6-serrotrehpotsirhc/pyproject.toml
[tool.poetry]
name = "cmsc408-sp2025-hw6"
version = "0.1.0"
description = ""
authors = ["John Leonard <jdleonard@vcu.edu>"]
package-mode = false
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<=3.13.6"
matplotlib = "^3.8.0"
plotly = "^5.18.0"
jupyter = "^1.0.0"
tabulate = "^0.9.0"
markdown = "^3.5"
pandas = "^2.1.1"
sqlalchemy = "^2.0.22"
pymysql = "^1.1.0"
mysql-connector-python = "8.3.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
